Add automatic display brightness based on ambient light level
This adds code to read the LDR and set brightness smoothly. This may be completely
overkill, but I'll see how it performs with real use. The problem could be solved
with less program space using a single threshold with hysteresis to switch between
"day" and "night" modes.

I suspect flickering brightness might be an issue with this, which means it would
need hysteresis anyway. With this chnage the code is now 966 bytes, so something
might need to be compacted or removed to make room for the timezone button handling.