Try to avoid timezone changing occassionally without button press
The timezone has skipped forward 1-2 hours a few times in the last month.
Something (maybe power glitches while it's already very dark) is causing
the ADC reading to drop below the button-press threshold for one reading,
which is advancing the timezone.

This changes the logic to advance the timezone at the end of the 500ms
period, requiring 5 readings 100ms apart to register as button presses.
This should stop any momentary glitches in power or the ADC from causing
an unwanted timezone change.