Move time mutation out of display_update() for grok-ability
Coming back to this code after 9 months, I was irked by the display_update()
mutating the time each call. This isn't technically a problem as the
function is only called once after reading the time from the GPS, but
it seems clearer to separate the application of timezone offset to the
updating of digits on the display.

This change doesn't affect binary size.