Replace AVR eeprom read/write functions to save 36 bytes program space
This uses the EEPROM read and write implementations from the datasheet,
but strips out waits for the previous write to complete. This check isn't
needed as our reads only occur on start-up, and writes are at least a
few seconds apart due to the way the timezone changing UI works.

Defining these statically lets the compiler optimise further as these
functions are only called once each.