Compile as a single unit to maximise compiler optimisation opportunities
This saves 84 bytes in the current code at the cost of making the compiled
assembly harder to follow as 85% of the program becomes a single function.

Since space saving is the name of the game, this is a worthwhile compromise.

Introduces a define AVRSTATIC so functions can be made static when
compiling the firmware, and left as `auto` when compiling the test
cases. If this ends up being annoying or confusing, the test cases can
just be changed to import source files too and the headers can be
removed.