Fix display intensity wrapping to 0x10 instead of maxing out at 0x0F
The display brightness value was able to go to 16, which was interpreted
by the MAX7219 as the lowest brightness (as it only looks at the low nibble).
Fix this so it can go brighter than the number of brightness table entries.

This didn't occur during normal use. I only noticed this when shining a
bright light directly at the LDR while testing.